Reorganize the test files
